Fixes #245 - applied by pos - thanks merkur2k

+++ src/new-installer/mce-install.sh
@@ -9,8 +9,7 @@ ntpdate ntp.ubuntu.com
 
 if [ ! -e /tmp/mce_wizard_data.sh ]
 then
-#	cp ./mce_wizard_data.sh /tmp
-	if [ `ifconfig -s | awk '$1 != "Iface" && $1 != "lo" { print $1 }' | wc -l` > 1 ]
+	if  [[ `ifconfig -s -a  | awk '$1 != "Iface" && $1 != "lo" && $1 != "pan0" { print $1 }' | wc -l` > 1 ]]
 	then
 		cp ./mce_wizard_data-double_nic.sh /tmp/mce_wizard_data.sh
 		#Use these for the defaults if we cannot automatically determine which to use
@@ -19,27 +18,36 @@ then
 		intif="eth1"
 		if route -n | grep -q '^0.0.0.0'
 		then
-		  #We have a default route, use it for finding external interface.
+			#We have a default route, use it for finding external interface.
 			extif=`route -n | awk '$1 == "0.0.0.0" { print $8 }'`
 			#Use the first available interface as the internal interface.
-		  for if in `ifconfig -s | awk '$1 != "Iface" && $1 != "lo" { print $1 }'`
-		  do
-			 if [ "$if" != "$extif" ]
-			 then
-				  intif=$if
-				  break
-			 fi
-		  done
+		  	for if in `ifconfig -s -a | awk '$1 != "Iface" && $1 != "lo"  && $1 != "pan0" { print $1 }'`
+		  	do
+				if [ "$if" != "$extif" ]
+			 	then
+					intif=$if
+					break
+			 	fi
+		  	done
 		fi
-		echo "Using $extif for external interface."
+		echo "Using $extif for external interface"
 		echo "Using $intif for internal interface"
 		
-		sed -e "s,\({extif}\),$extif,g" /tmp/mce_wizard_data.sh > /tmp/tmpfile && mv /tmp/tmpfile /tmp/mce_wizard_data.sh
-		sed -e "s,\({intif}\),$intif,g" /tmp/mce_wizard_data.sh > /tmp/tmpfile && mv /tmp/tmpfile /tmp/mce_wizard_data.sh
+		sed --in-place -e "s,\({extif}\),$extif,g" /tmp/mce_wizard_data.sh
+		sed --in-place -e "s,\({intif}\),$intif,g" /tmp/mce_wizard_data.sh
 	else
+		extif=eth0
+		if route -n | grep -q '^0.0.0.0'
+                then
+                	#We have a default route, use it for finding external interface.
+                	extif=`route -n | awk '$1 == "0.0.0.0" { print $8 }'`
+		fi
 		cp ./mce_wizard_data-single_nic.sh /tmp/mce_wizard_data.sh
+		echo "Using $extif for single nic install"
+		sed --in-place -e "s,\({extif}\),$extif,g" /tmp/mce_wizard_data.sh
 	fi
 fi
+
 rm -rf /tmp/mce_installer_error
 InstallerLogFile="/var/log/mce-installer-$(date +%Y%m%d_%H%M%S).log"
 exec &> >(tee $InstallerLogFile)
@@ -52,7 +60,6 @@ if [[ ! -r /tmp/mce_wizard_data.sh ]] ;then
         ExitInstaller "Wizard Information is corrupted or missing."
 fi
 . /tmp/mce_wizard_data.sh
-
 
 Core_PK_Device="0"
